+++ R/dateSequence.R
@@ -86,7 +86,7 @@ dateSequence <- function(
   stopIfNull(ceilingEnd)
 
   if ( !timezone %in% base::OlsonNames() )
-    stop(paste0("Timezone '", timezone, "' is not recognized."))
+    stop(sprintf("'timezone = %s' is not found in OlsonNames()", timezone))
 
   # ----- Create sequence ------------------------------------------------------